+++ fuse/op_field.c
@@ -104,7 +104,7 @@ field_open(struct bibfs_state *b,
 
 static int
 field_read(struct bibfs_state *b,
-	char *buf, size_t size, off_t offset, struct fuse_file_info *fi,
+	char *buf, size_t size, size_t offset, struct fuse_file_info *fi,
 	const char *key, const char *name, const char *ext)
 {
 	struct bib_entry *e;
