+++ fuse/op_index.c
@@ -84,7 +84,7 @@ done:
 
 static int
 index_read(struct bibfs_state *b,
-	char *buf, size_t size, off_t offset, struct fuse_file_info *fi,
+	char *buf, size_t size, size_t offset, struct fuse_file_info *fi,
 	const char *key, const char *name, const char *ext)
 {
 	struct bib_entry *e;
